Bootstrap Implementation
Added and installed the Bootstrap + Bootstrap-React packages to the code which imported the bootstrap libraries.

Changes were made to the Header. Swapped out the default CSS buttons to use the Bootstrap buttons. The header section elements were also put into a container to test out positioning and spacing of certain elements. The home button is disabled to resemble if the user was already on the home page.

Changes were made to the Footer. Swapped out the default CSS buttons to use the Bootstrap buttons. The home button is disabled to resemble if the user was already on the home page.