RTTM v0.2.2
Many changes were made to the website! A large of the code was changed or reorganized. The main page has been worked on currently displaying the text box and a AWS Video Player. The video player is functional for purposes with IVS but is non-functional for purposes needed for this project due to a change with the system in charge of gathering and handling the video. The header and footers have been adjusted to be able to fix a bug that prevented it from covering the entire width of the browser window. In the footer, links were added to replace the buttons that were initially added. A gradient background was added! Currently doesn't fill up the entire height of the window screen but is set to do so when more content is added. The base foundation was implemented to be able to use React Router v6 and add more pages to the website!