RTTM v0.3
Various changes were made to the entire website! An npm re-install may be necessary to be able to view the site properly.

The main page has been redesigned to follow the new changes made on the RTTM Website Design (https://xd.adobe.com/view/d1945576-2f88-4dd4-b39a-2986171a9784-84c7/). The layout of the main page is set up to display 2 video players on the top, followed by charts that correspond to the video feed down below. The current charts on this version are only for demonstration purposes and do not correlate with what would be displayed on the video player.

A working hls video player has been added to the website! This is currently connected to the bucket Alex has set up and is functional when he runs his system. The video player on the left side is functional while the video player on the right is not and is currently only for display purposes until we can get a second camera functionality set up.

The technologies page has been updated to display some of the information involving what the website and project is about and the technologies involved are.

Changes and fixes were done to the Header & Footer
> The header and footer have been updated! The header color has been changed to correlate to the new website design and color scheme, and the links have been updated to be centered with the header overall. The footer has been updated to now have the links centered with the footer overall. Links on both the header and the footer are now all functional.

For convenience, many files have been moved around and organized into folders that correspond to their function or their pages for better convenience.

Graphs on the main website are still being worked on as of this version. Test templates are what is currently being displayed.