Video Adjustments and CSS changes
Changed the video that is played on the mp4 player. Also adjusted the player to autoplay the video, loop the video, and disable user controls.

Video boxes and players were adjusted to increase in height. Player 1 height now at 413 and video-box height now at 26.2em

Header was adjusted with a set size of 70pt